Koncat.Core

Description

Koncat.Core exports the bootstrap function, which is used to start the Bot. Internally, Koncat.Core handles connecting to the server and dispatching server queries.

Starting the Bot

bootstrap :: IO ()

Set up Exception handling, start off the conversation with the server and start providing IRC Bot functionalities.