Koncat.Core.API

Description

Koncat.Core.API exports functions that increase the abstraction level of IRC, making working with it more comfortable.

Currently, this Module is also temporarily acting as a configuration file. Configuration will get its own File in the near future.

Documentation

module Control.Monad.IO.Class

Configuration

debug :: Bool

Specify if Debug messages should be displayed

server :: String

The server the Bot should connect to

port :: String

The port the Bot should connect to

nick :: ByteString

The Nickname of the Bot

chan :: [Channel]

The Channel the Bot should join

dbgChan :: Channel

The Channel Debug-Messages are posted to.

Low-level Interfacing

r_write :: ByteString -> IRC ()

Write a raw String to the Server, also logging to stdout

r_write' :: Handle -> ByteString -> IO ()

Same than r_write, except for the function not being inside the IRC Monad thus requiring the Handle to be passed as an argument.

Events

namely :: ByteString -> ByteString -> Bool

Event based on a match of (the beginning of) a Line coming from the Server

onServerCode :: ByteString -> ByteString -> Bool

Event based on a Control Code coming from the Server

onUserAction :: ByteString -> ByteString -> Bool

Event based on Action coming from a user

API Functions

joinChannel :: Channel -> ByteString -> IRC ()

Join a Channel

partChannel :: Channel -> IRC ()

Leave a Channel

joinAll :: [Channel] -> ByteString -> IRC ()

Join multiple channels

retPing :: ByteString -> IRC ()

Respond to a Ping

recvArgs :: ByteString -> ByteString

Get the actual message from a Client, or the Server

privmsg :: Channel -> ByteString -> IRC ()

Send a message to the channel

privmsg' :: Handle -> Channel -> ByteString -> IO ()

See r_write'

incCounter :: IRC ()

Increment the Counter the State is Carrying

Convenience

io :: IO a -> IRC a

Lift IO in the IRC Monad

Miscellaneous

hasFlag :: String -> Flags -> Bool

Check if a flag is set, discarding eventual values bound to it.

getFlag :: String -> Flags -> Maybe String

Check if a flag is set and return the value of it if set.

isPrivateMsg :: GenericIRC -> Bool

Check if a message someone is sending is private.

replace :: Eq a => [a] -> [a] -> [a] -> [a]

Find/Replace something

printf' :: PrintfType r => ByteString -> r

printf a ByteString