Koncat.Core.Native.Permissions

Description

This Module implements the Permission-System of the bot.

Maintenance of User Information is done in Koncat.Permissions. Koncat.Core.Native.Permissions is providing the actual bindings to IRC that users can use to login, logout and register.

Todo:

grantUserPermissions function

Logout user when IRC is quit

Cross-Check with Permission-System of the IRC Server in order to confirm the identity of a user.

Documentation

updateUserDB_userQuit :: ByteString -> IRC ()

The user has quit, so log them out from the system.

updateUserDB_identQuery :: GenericIRC -> IRC ()

A user has requested identification.

Verify their credentials and log them in, or display an error message if the credentials are wrong.

handleRegister :: GenericIRC -> IRC ()

A user wants to register.