Koncat.Module

Description

Koncat.Module is responsible for starting and stopping Modules, checking if the Module isn't annoying the IRC Server (for example flooding it, which could also lead into the server killing our bot) and providing the Modules with the Koncat API

Documentation

execMod :: ModuleCall -> IRC ()

Execute a Module, streaming its output into IRC.

constructModuleCall

Arguments

:: (GenericIRC, IRCCommand)
Information about the user, information about the command the user issued

-> String
the modulePath, as retrieved via Koncat.Source.expose

-> ModuleCall

Collect Information neccessary to start a Module.