Koncat.Parsers

Description

Koncat.Parsers exports various parsers neccessary to efficiently handle user queries. It is using the Parsec Text Parsing library internally.

Todo:

Configuration File instead of defaultValues

Search for cases where parseIRC might fail, rewrite

Parsing Module Handshakes

parseHandshake :: ByteString -> Handshake

parseHandshake gets the raw text as recieved when asking a module for a handshake, parses it and stuffs it in the Handshake Data Type.

parseAlias :: ByteString -> (String, String)

Extract 2 quoted Strings from one String.

>>> parseAlias "\"First String\" \"Second String\""
("First String", "Second String")

runFuncParser :: ByteString -> [(String, String, String, Maybe [(Char, String)])]

runFuncParser will parse the function part of the Handshake. See Figure 1 (-> Source Code) for further elaboration.

>>> runFuncParser "{helloworld:[A Hello World Function](*String)%E=5s%P=Admin}"
[("helloworld","A Hello World Function","*String",Just [('E',"5s"),('P',"Admin")])]

Default values for Arguments

defaultValues :: [(Char, String)]

Hardcoded lookup-table that describes default values for variables. They will be inserted in the SQL Database if the user didn't set it!

Parsing IRC Messages

parseIRC

Arguments

:: ByteString
Raw Input String

-> GenericIRC
Parsed IRC Message

parseIRC parses everything except the actual message the Bot recieves from IRC, and stuffs it in the GenericIRC type.

>>> parseIRC ":tsv!tsv@localhost PRIVMSG #main :This is the actual message "
GenericIRC {user = UserInfo {name = "tsv", rName = "tsv", vhost = "localhost"}, orig = "#main", msg = ["This","is","the","actual","message"]}

See Figure 2 (Source Code) for further reference on implementation

See RFC 1459, http://www.faqs.org/rfcs/rfc1459.html for further reference on the IRC specification.

parseIRCFunctionCall :: ByteString -> Either ParseError IRCCommand

parseIRCFunctionCall is parsing an IRC message which is directed to the Bot and referring to a Module.

>>> parseIRCFunctionCall ":util.testytest #dothingsheh(Arg1,Arg2) -flag -flag2=SomeArgument"
Right (IRCCommand {qModule = "util.testytest", qFunc = "dothingsheh", qArgs = Just ["Arg1","Arg2"], qFlags = Just [("flag",Nothing),("flag2",Just "SomeArgument")]})

Parsing (execution-)Time

getTime :: ByteString -> Int

Get Microseconds from int{s|sec|min|h|d}, as required for setting a timeout using System.Timout

>>> getTime "10s"
10000000

>>> getTime "2min"
120000000