Koncat.Permissions

Description

Various functions neccessary to control Module Execution by implementing Permissions.

Documentation

registerUser

Arguments

:: ByteString
Username

-> ByteString
Password

-> IO ()

Register a user

rebuildPermDB :: IO ()

Set up the database structure

loginUser :: ByteString -> ByteString -> IO (Maybe Permission)

Login a user

toPermission :: String -> Permission

Convert a String to a Permission

TODO:

Simply make an instance of Read

userExists :: ByteString -> IO Bool

Check if a user exists in the Database

getPermission :: IRCCommand -> IO Permission

Query the database for a Permission of a user

unlessHasPrivilege :: Nick -> IRCCommand -> IRC () -> IRC () -> IRC ()

If user has enough privileges, do something. Otherwise, do something else.