Koncat.PrettyPrint

Description

Koncat.PrettyPrint is providing a few option functions that make Bot-Output more aesthetically pleasing for the end user. Note that Syntax Highlighting is done in a different Module, that being Koncat.Language.SyntaxHighlighter

Documentation

drawTable' :: [[ByteString]] -> ByteString -> [ByteString]

Align a Multi-Dimensional Array

>>> putStrLn $ drawTable [["thisIsRow1","AnotherElemInRow1","alsoInRow1"],["thisIsRow2","moreText","EvenMoreText"]] " | "
thisIsRow1        | thisIsRow2   | 
AnotherElemInRow1 | moreText     | 
alsoInRow1        | EvenMoreText | 

However, it may be more convenient for you to input text per column:

>>> putStrLn $ drawTable (transpose [["column_one","column_two","column_three"],["alsoFirstcolumn","AndTheresTheSecondcol", "...."]]) " | "
column_one      | column_two            | column_three | 
alsoFirstcolumn | AndTheresTheSecondcol | ....         | 

where transpose is from Data.List

See drawTable. The final result isn't concatentated.