Koncat.Source

Description

Koncat.Source will source child directories of mod/ for modules. It will also expose a few functions for working with the .koncat.cache.db, a database where temporary Module Information is stored and constantly rebuilt.

Plugins will be available to the user by calling it with its (mod-)path from IRC.

Implementation:

Every Plugin, no matter the language it is written in, has to implement a handshake, exposing functions, description for functions and various other metadata to the Bot.

For elaboration on the Syntax of handshakes, check Koncat.Parsers

This Module proceeds to store the information provided by the modules in a SQLite database.

The user then is able to call a Function of a Module from IRC using the Syntax described above.

Notes:

Functions of modules are called by printing their name in the modules STDIN.

A Directory may be excluded from sourcing by putting a file named .dont.src in it.

General functions

rebuild :: IO ()

Rebuild the database

rmDB :: IO ()

Remove the Database

Operations on modules

update :: FilePath -> IO ()

Rebuild the cache of a specific module. This function uses the filename, not the modname.

expose :: String -> IO (Maybe FilePath)

Search the database if there is a module matching a modname. Returns the Path to the module, or Nothing.

exposeMeta :: String -> IO (Maybe [[ByteString]])

Search the database and return the (name,args, permissions, desc) of the functions.

dropMod :: FilePath -> IO ()

Remove a module from the database. This function uses the filename, not the modname.

getMaxExecTime :: FilePath -> String -> IO Int

Return the maximum execution time of a Function in microseconds, given the filepath and the function

fetchModules :: IO [ByteString]

List modules that are available.

safeMod

Arguments

:: String
Modname, for example "util.test"

-> String
Query, for example "SELECT * FROM {DB}". {DB} is a placeholder for the modname.

-> [SqlValue]
Additional values to be inserted in the database.

-> IO (Maybe [[SqlValue]])

When working with [mod_some.module] tables, doing something like

 SELECT * FROM [mod_" ++ getModName someModule ++ "]"

will result in a SQL Injection vulnerability.

This function will safely execute queries referring to such tables.

>>> safeMod "util.test" "SELECT * FROM {DB}" []
Just [[SqlByteString "helloworld",SqlByteString ....

>>> safeMod "inputContainingSqli" "SELECT * FROM {DB}" []
Nothing

True/False queries on the database

doesTableExist :: String -> IO Bool

Check if a table in the database exists.

funcExists :: IRCCommand -> IO Bool

Check if the functioncall of a module is vaild.

doesModuleExist :: String -> IO Bool

Check if a module exists