Koncat.Types

Description

Various Types.

Documentation

type Nick = ByteString

The Nickname of a user

type Channel = ByteString

An IRC channel

type Flags = Maybe [(String, Maybe String)]

Flags the User is providing to the Bot

type IRC = StateT Bot IO

Our IRC Monad

data ArgT

A single argument as supplied to a function.

Constructors

Varidic
The function can take any number of arguments

Void
The function takes no arguments

String
The function takes a string as an argument

Int
The function takes a (32-Bit) Integer as an argument

Double
The function takes a (64-Bit) double as an argument

data Permission

Represents the Permission that is neccessary to call a Module i.e the permission the Module is requesting the user to have.

Note that this data differs from the levels of permission someone is able to assume.

Constructors

Admin
Having the highest possible priority, Admins are always able to execute a Module.

ANY
Next is ANY, i.e the Module is not requesting the user to have any special permissions.

Privileged
Only privileged users and above are allowed to execute

Registered
Only registered users and above are allowed to execute

UnRegistered
Having the lowest possible priority, UnRegistered is only able to execute if ANY or UnRegistered is set.

data Bot

Bot is carrying all the data we need in order to dispatch queries.

Constructors

Bot

Fields

iSock :: Handle

The Socket that's connected to the server.

starttime :: ClockTime

The starttime, for uptime tracking (not yet implemented)

modCnt :: Integer

A counter containing the number of Modules tracked

aliasDB :: [(ByteString, ByteString)]

Aliases that are set

userDB :: Map Nick Permission

Users that have Registered

channelDB :: [Channel]

Channels the Bot joined

data GenericIRC

GenericIRC holds information contained in a single line of input coming from the server. (see parseIRC)

Constructors

GenericIRC

Fields

user :: UserInfo

Information about the User

orig :: Channel

Where the Message is coming from. Note this may also be a User, not only a Channel.

msg :: [ByteString]

The actual Message.

data UserInfo

A minimal amount of Information about the User a Message is coming from

Constructors

UserInfo

Fields

name :: Nick

The Nickname of the user

rName :: Nick

the RealName of the user

vhost :: ByteString

the vhost of the user

data Handshake

Information a Module sends over during the HELO-Handshake

Constructors

Handshake

Fields

desc :: ByteString

The Description of the Module

funcs :: [(String, String, String, Maybe [(Char, String)])]

Functions the Module sends over (see Koncat.Parsers)

author :: ByteString

The Auther of the Module.

data IRCCommand

A message coming from the user directed to the Bot (not referring to a native call)

Constructors

IRCCommand

Fields

qModule :: String

the Module the user attempts to call

qFunc :: String

the Function the user attempts to call

qArgs :: Maybe [String]

Arguments the user is delivering to the function

qFlags :: Flags

Flags the user is transmitting to the Bot

data ModuleCall

Information that is neccessary in order to start a module

Constructors

ModuleCall

Fields

path :: FilePath

The path of the Module

func :: String

The function of the Module to be called

args :: Maybe [String]

Arguments the function should be called with

dest :: Channel

Where output should be redirected to (may also be a user)

caller :: Nick

The Nick who called the Module (TODO: make it UserInfo)

flags :: Flags

Flags